% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_mean.R
\name{beta_mean}
\alias{beta_mean}
\title{Caculate the mean of beta of many cond files and return all beta in your cond
files.}
\usage{
beta_mean(cond_name = NA)
}
\arguments{
\item{cond_name}{Must be a charactor that you need to analysis.}
}
\value{
A list contain the mean of beta and all beta value.
}
\description{
\code{beta_mean} return a list which is mean of beta in some cond files that
in one experiment condition. It also return all beta value in choosen files.
}
\author{
Xiaokai Xia (夏骁凯)
\email{xia@xiaokai.me}
}
