% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_spm_fnirs.R
\name{convert_spm_fnirs}
\alias{convert_spm_fnirs}
\title{Convert file generated by Labnirs to the file spm_fnirs need}
\usage{
convert_spm_fnirs(filename = NA)
}
\arguments{
\item{filename}{Is path to locate a Labnirs file.}
}
\value{
Some csv file you need use them in spm_fnirs.
}
\description{
Convert the data file generated by Labnirs to
the spm_fnirs toolbox in Matlab.
}
\details{
This function can generate some file which spm_fnirs need. You should choose
the file generated by Labnirs, then this program will generate some file at
the same folder. "wavelength1"and "wavelength2" are two csv file which
spm_fnirs need, and you should know the wavelength 1 is 780nm and wavelength
2 is 805nm. We also get onset files, and marked the onset as seconds or line
number in Labnirs file. The last file is ch_config, spm_fnirs need it to know
the pair of probe.
}
\author{
Xiaokai Xia (夏骁凯)
\email{xia@xiaokai.me}
}
