% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fNIRS_beta.R
\name{extract_fNIRS_beta}
\alias{extract_fNIRS_beta}
\title{Extract beta value in con files}
\usage{
extract_fNIRS_beta(filename = NA)
}
\arguments{
\item{filename}{Is a path of single con file you need to extract the beta
value in it.}
}
\description{
\code{extract_fNIRS_beta} can extract the beta value in con file generated by
spm_fnirs.
}
\author{
Xiaokai Xia \email{xia@xiaokai.me}
}
