% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_map.R
\name{layout_map}
\alias{layout_map}
\title{Generate the layout map for Labnir}
\usage{
layout_map(ncol = NA, nrow = NA, save_filename = "layout.png")
}
\arguments{
\item{ncol}{Is the column number in your layout}

\item{nrow}{Is the row number in your layout}

\item{save_filename}{Is the filename for your layout picture}
}
\description{
Generate a layout map which you will need it in your paper
}
\author{
Xiaokai Xia (夏骁凯)
\email{xia@xiaokai.me}
}
