% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nii_heatmap.R
\name{nii_heatmap}
\alias{nii_heatmap}
\title{Map the nii file as standard heatmap by ggplot2}
\usage{
nii_heatmap(nii_image)
}
\arguments{
\item{nii_image}{Must be a single nii image data. If your nii data is a list
please use \code{\link{mean_group_nii}} to get a single image data.}
}
\value{
An picture plot by ggplot2.
}
\description{
\code{nii_heatmap} can map nii file as standard heatmap by using ggplot2. So
you will get a legend to tell you the meaning of colour in this heatmap.
}
\author{
Xiaokai Xia (夏骁凯) \email{xia@xiaokai.me}
}
